+++ NEWS.md
@@ -17,6 +17,7 @@
 - Add a new `format` trigger to invalidate targets when the specialized data format changes (#1104, @kendonB).
 - Add new functions `cache_planned()` and `cache_unplanned()` to help selectively clean workflows with dynamic targets (#1110, @kendonB).
 - Add S3 classes and pretty print methods for `drake_config()` objects and `analyze_code()` objects.
+- Add a new `"qs"` format (#1121, @kendonB).
 
 ## Speedups